Wrote tests for Softmax and MSELoss
