+++ core/CML/Functions/Loss.h
@@ -1,0 +1,6 @@
+#ifndef __CML_FUNCTIONS_LOSS_H__
+#define __CML_FUNCTIONS_LOSS_H__
+
+#include "Loss/MSELoss.h"
+
+#endif // __CML_FUNCTIONS_LOSS_H__
