+++ core/CML/Functions/Loss/MSELoss.h
@@ -16,12 +16,12 @@ namespace Function {
             switch(reduction){
                 case nn::Reduction::MEAN:
                     t = make_tensor<T>(static_cast<DMatrix<T>>(
-                        (actual->data() - expected->data()).array().square().rowwise().mean()
+                        (actual->data() - expected->data()).array().square().colwise().mean()
                     ));
                     break;
                 case nn::Reduction::SUM:
                     t = make_tensor<T>(static_cast<DMatrix<T>>(
-                        (actual->data() - expected->data()).array().square().rowwise().sum()
+                        (actual->data() - expected->data()).array().square().colwise().sum()
                     ));
                     break;
                 default:
@@ -31,15 +31,35 @@ namespace Function {
                     break;
             }
             t->computeGrad = true;
-            // t->initGraph({actual, expected}, [reduction](std::vector<tensor<T>>& params, std::vector<tensor<T>> output) -> std::vector<tensor<T>> {
-            //     std::cout << "MSELoss.actual:" << std::endl << actual << std::endl;
-            //     std::cout << "MSELoss.expected:" << std::endl << expected << std::endl;
-            //     auto u = make_tensor<T>(static_cast<DMatrix<T>>(
-            //         2*(actual->data() - expected->data())
-            //     ));
-            //     std::cout << "MSELoss.u:" << std::endl << u << std::endl;
-            //     return u;
-            // });
+            t->initGraph({actual, expected}, [reduction](std::vector<tensor<T>>& params, std::vector<tensor<T>> output) -> std::vector<tensor<T>> {
+                auto actual = params.at(0);
+                auto expected = params.at(1);
+
+                T c = 0;
+                switch(reduction){
+                    case nn::Reduction::MEAN:
+                        c = (T)(2.0/actual->size());
+                        break;
+                    case nn::Reduction::SUM:
+                        c = (T)(2);
+                        break;
+                    default:
+                        c = (T)(2);
+                        break;
+                }
+
+                tensor<T> actual_grad = make_tensor<T>(static_cast<DMatrix<T>>(
+                    c*(actual->data() - expected->data())
+                ));
+#ifdef DEBUG
+                using namespace std;
+                cout << "MSELoss::actual = " << actual << endl;
+                cout << "MSELoss::expected = " << expected << endl;
+                cout << "MSELoss::actual_grad = " << actual_grad << endl;
+#endif
+
+                return {actual_grad, nullptr};
+            });
             return t;
         }
